requirements.txt
faiss_cpu==1.7.2
fastapi==0.98.0
numpy==1.23.5
Pillow==9.5.0
starlette==0.28.0
torch==2.0.1
transformers==4.30.0
uvicorn==0.22.0
